fix(statistics): Fix render issues with agg trip data

+++ internal/models/timespace_aggregation.go
@@ -2,6 +2,6 @@ package models
 
 
 type TimeSpaceAggregation struct {
-	TotalHours int `json:"total_hours"`
+	TotalHours float32 `json:"total_hours"`
 	TotalKm int `json:"total_km"` // Total kilometers traveled
 }
